Restructured profile tabs
