+++ app/src/main/res/layout/post_fragment.xml
@@ -61,9 +61,9 @@
                 android:padding="4dp"
                 sparkbutton:activeImage="@drawable/ic_bookmark_black_24dp"
                 sparkbutton:iconSize="28dp"
-                sparkbutton:inactiveImage="@drawable/ic_bookmark_black_24dp"
-                sparkbutton:primaryColor="@color/share_blue"
-                sparkbutton:secondaryColor="@color/black"
+                sparkbutton:inactiveImage="@drawable/ic_bookmark_border_black_24dp"
+                sparkbutton:primaryColor="@color/black"
+                sparkbutton:secondaryColor="@color/share_blue"
                 app:layout_constraintBottom_toBottomOf="@+id/profilePic"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintTop_toTopOf="@+id/profilePic" />
